requirement.txt
asgiref==3.7.2
bcrypt==4.0.1
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
Django==4.2.6
django-cors-headers==4.3.0
django-environ==0.11.2
djangorestframework==3.14.0
idna==3.6
Pillow==10.1.0
PyJWT==1.7.1
pypng==0.20220715.0
pytz==2023.3.post1
qrcode==7.4.2
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
